Split account HTTP routes
